requirements.txt.txt
streamlit
pandas
plotly
fpdf2
arabic-reshaper
python-bidi